requirements.txt
streamlit==1.37.0
python-dotenv

sqlalchemy>=2.0
pydantic==1.10.13

# ML + Inference
ultralytics
torch==2.2.0+cpu
torchvision==0.17.0+cpu
torchaudio==2.2.0+cpu
--extra-index-url https://download.pytorch.org/whl/cpu

opencv-python
pillow

# API + Utilities
openai
requests
tiktoken

# General utilities
numpy
pandas
streamlit-searchbox
streamlit-drawable-canvas==0.9.3
bcrypt
